Allow decimal input

+++ main.cpp
@@ -11,7 +11,7 @@ int main(int argc, char **argv)
         std::cerr << "Usage: main.exe -t <number_of_tokens>" << endl;
         return 1;
     }
-    int numTokens = atoi(argv[2]);
+    double numTokens = atof(argv[2]);
     if (numTokens <= 0)
     {
         std::cerr << "Invalid number of tokens" << endl;
